requirements.txt
insightface==0.7.3
onnxruntime-gpu>=1.19.2,<2.0
opencv-python-headless>=4.9.0
numpy>=1.26.4,<2.0
Pillow>=10.3.0
pillow-heif>=0.16.0
scipy>=1.12.0,<1.13
fastapi>=0.111.0
uvicorn>=0.31.1
python-multipart>=0.0.9